wip on malloc

+++ sys/sys/malloc.h
@@ -1,2 +1,2 @@
 void malloc_init(void);
-void malloc_add_region(int base, int size);
+void malloc_add_region(uintptr_t base, int size);
